Специальности перенесены к Админу. Группы перенесены в ученики и созданы полиси для функций

+++ app/Http/Controllers/GroupController.php
@@ -63,23 +63,35 @@ class GroupController extends Controller
         }
 
         //Специальности
-        if(isset($teacher->city_id)){
-            $specialities = Speciality::where("city_id", $teacher->city_id)
-                ->orderBy('id', 'asc')
-                ->get();
-        }else{
-            if($request->user()->hasRole('teacher')){
-                $admin_id = $teacher->teacher_admins[0]->id;
-                $admin = User::find($admin_id);
-            }
-            if($request->user()->hasRole('administrator')){
-                $admin = Auth::user();
-            }
-            $specialities = $admin->specialities()
-                ->where("city_id", null)
-                ->orderBy('id', 'asc')
-                ->get();
+        // if(isset($teacher->city_id)){
+        //     $specialities = Speciality::where("city_id", $teacher->city_id)
+        //         ->orderBy('id', 'asc')
+        //         ->get();
+        // }else{
+        //     if($request->user()->hasRole('teacher')){
+        //         $admin_id = $teacher->teacher_admins[0]->id;
+        //         $admin = User::find($admin_id);
+        //     }
+        //     if($request->user()->hasRole('administrator')){
+        //         $admin = Auth::user();
+        //     }
+        //     $specialities = $admin->specialities()
+        //         ->where("city_id", null)
+        //         ->orderBy('id', 'asc')
+        //         ->get();
+        // }
+
+        if($request->user()->hasRole('teacher')){
+            $admin_id = $teacher->teacher_admins[0]->id;
+            $admin = User::find($admin_id);
         }
+        if($request->user()->hasRole('administrator')){
+            $admin = Auth::user();
+        }
+        $specialities = $admin->specialities()
+            ->where("city_id", null)
+            ->orderBy('id', 'asc')
+            ->get();
 
 
         return view('teacher.group_index', [
@@ -160,6 +172,7 @@ class GroupController extends Controller
      */
     public function edit(Request $request, Group $group)
     {
+        $this->authorize('groupProtected', $group);
 
         $teacher = $request->user();
 
@@ -183,30 +196,39 @@ class GroupController extends Controller
                 ->where("city_id", null)
                 ->orderBy('id', 'asc')
                 ->get();
-
-            // foreach($students as $student){
-            //     dd($student->groups->pivot->group_id);
-            // }
         }
 
         //Специальности
-        if(isset($teacher->city_id)){
-            $specialities = Speciality::where("city_id", $teacher->city_id)
-                ->orderBy('id', 'asc')
-                ->get();
-        }else{
-            if($request->user()->hasRole('teacher')){
-                $admin_id = $teacher->teacher_admins[0]->id;
-                $admin = User::find($admin_id);
-            }
-            if($request->user()->hasRole('administrator')){
-                $admin = Auth::user();
-            }
-            $specialities = $admin->specialities()
-                ->where("city_id", null)
-                ->orderBy('id', 'asc')
-                ->get();
+        // if(isset($teacher->city_id)){
+        //     $specialities = Speciality::where("city_id", $teacher->city_id)
+        //         ->orderBy('id', 'asc')
+        //         ->get();
+        // }else{
+        //     if($request->user()->hasRole('teacher')){
+        //         $admin_id = $teacher->teacher_admins[0]->id;
+        //         $admin = User::find($admin_id);
+        //     }
+        //     if($request->user()->hasRole('administrator')){
+        //         $admin = Auth::user();
+        //     }
+        //     $specialities = $admin->specialities()
+        //         ->where("city_id", null)
+        //         ->orderBy('id', 'asc')
+        //         ->get();
+        // }
+
+        if($request->user()->hasRole('teacher')){
+            $admin_id = $teacher->teacher_admins[0]->id;
+            $admin = User::find($admin_id);
         }
+        if($request->user()->hasRole('administrator')){
+            $admin = Auth::user();
+        }
+        $specialities = $admin->specialities()
+            ->where("city_id", null)
+            ->orderBy('id', 'asc')
+            ->get();
+
 
         return view('teacher.group_edit', [
             'group' => $group,
@@ -224,6 +246,8 @@ class GroupController extends Controller
      */
     public function update(Request $request, Group $group)
     {
+        $this->authorize('groupProtected', $group);
+
         $validator = Validator::make($request->all(), [
             'title' => 'required|max:255',
             'students_id' => 'required',
@@ -235,16 +259,53 @@ class GroupController extends Controller
               ->withErrors($validator);
         }
 
-        dd("сделать детач всех пользователей этой группы или сделать проверку по существующим связям");
+        // dd("сделать детач всех пользователей этой группы или сделать проверку по существующим связям");
 
         $groupUpdated = $group->update([
             'title' => $request->title,
             'speciality_id' => $request->speciality_id ? $request->speciality_id : null,
         ]);
 
-        // foreach($request->students_id as $student_id){
-        //     $groupUpdated->students()->attach($student_id, ['group_id' => $groupUpdated->id, 'student_id' => $student_id]);
-        // }
+
+        $oldStudents = [];
+        foreach($group->students as $studenInGroup){
+            array_push($oldStudents, $studenInGroup->id);
+        }
+        // dd($oldStudents);
+
+        $newStudents = [];
+        foreach($request->students_id as $student_id){
+            array_push($newStudents, (int)$student_id);
+        }
+        // dd($newStudents);
+
+        $detachStudents = [];
+        $attachStudents = [];
+        foreach($newStudents as $newStudent){
+            if($oldStudents != null){
+                foreach($oldStudents as $oldStudent){
+                    if(!in_array($newStudent, $oldStudents)){
+                        array_push($attachStudents, $newStudent);
+                    }
+                    if(!in_array($oldStudent, $newStudents)){
+                        array_push($detachStudents, $oldStudent);
+                    }
+                }
+            }else{
+                array_push($attachStudents, $newStudent);
+            }
+        }
+
+        $detachStudents = array_unique($detachStudents);
+        $attachStudents = array_unique($attachStudents);
+                                                    
+        foreach($detachStudents as $detachStudent_id){
+            $group->students()->detach($detachStudent_id);
+        }
+
+        foreach($attachStudents as $attachStudent_id){
+            $group->students()->attach($attachStudent_id, ['group_id' => $group->id, 'student_id' => $attachStudent_id]);
+        }
 
         Session::flash('message', 'Информация о группе успешно сохранена!');
 
@@ -257,9 +318,16 @@ class GroupController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Group $group)
     {
-        //
+        $this->authorize('groupProtected', $group);
+
+        $group->students()->detach();
+        $group->delete();
+
+        Session::flash('message', 'Группа успешно удалена и распущена');
+
+        return back();
     }
 
     /**
@@ -270,7 +338,10 @@ class GroupController extends Controller
      */
     public function ungroup(Group $group, User $student)
     {
+        $this->authorize('groupProtected', $group);
+
         $group->students()->detach($student->id);
+
         Session::flash('message', $student->second_name . ' '. $student->first_name . ' ' . $student->last_name . ' откреплен от группы ' . $group->title);
     
         return back();
